fix: further refactoring (#106)
* refactor(propsToState): remove eslint disable and add return type
* fix: remove wrong param from getCQPath
* fix: do not re-export constants
* fix: prepare withEditorContext for removal
* fix: add circular dependency check
* fix: get rid off circular dependency
