+++ src/constants/properties.ts
