+++ src/core/ComponentMapping.tsx
@@ -15,7 +15,7 @@ import React, { ComponentType } from 'react';
 import { ComponentMapping } from '@adobe/aem-spa-component-mapping';
 import { EditConfig } from './EditableComponent';
 import { ReloadableModelProperties } from './ModelProvider';
-import { withEditorContext } from '../EditorContext';
+import { withEditorContext } from '../delete/EditorContext';
 
 /**
  * @private
